requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-yasg==1.21.7
PyJWT==2.8.0
